==> 2.tratamento.Rmd <==
---
title: "Tratamento"
author: "Grupo 3 - Desafio 2"
date: "10/14/2020"
output: 
  html_document: 
    keep_md: yes
---
# Tratamento dos Dados

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# Pacotes necessários
library(tidyverse)
library(visdat)
library(VIM)
```

```{r leitura}
brasil_dados_basicos <- readRDS("brasil_dados_basicos.rds" )
brasil_movimentos <- readRDS("brasil_movimentos.rds" )
```

## Dados Básicos 

```{r}
str(brasil_dados_basicos)
```

```{r missings_db}
vis_miss(brasil_dados_basicos[4:6]) #CodigoAssunto, CodigoPaiNacional e Descricao 90% missing
vis_miss(brasil_dados_basicos[7:8]) # codigoNacional OK (só 9% missing)
vis_miss(brasil_dados_basicos[c(4, 8)]) #codigoNacional e codigoAssunto complementares (esperado)
vis_miss(brasil_dados_basicos[9:10]) # presentes mas não relevantes
vis_miss(brasil_dados_basicos[11:12]) # dadosBasicos.dataAjuizamento: ok, data em que o processo foi inicialmente recebido pelo órgao (entrada do processo)
vis_miss(brasil_dados_basicos[13:15]) # dadosBasicos.totalAssuntos e classeProcessual ok, nivelSigilo não relevante (todos não sigilosos)
vis_miss(brasil_dados_basicos[16:18]) # dadosBasicos.orgaoJulgador.nomeOrgao ok
vis_miss(brasil_dados_basicos[19:21]) # codigoOrgao ok instancia ok
vis_miss(brasil_dados_basicos[22:24]) # assuntoLocal praticamente tudo missing
vis_miss(brasil_dados_basicos[25:27]) # valorCausa 44% missing, competencia 54%, assuntoLocal 100%
vis_miss(brasil_dados_basicos[28:29]) # tamanhoProcesso e assuntoLocalPai praticamente tudo missing
```


```{r transformacao_db}
vars_int <- c("codigoAssunto", "codigoPaiNacional", "codigoNacional", "dadosBasicos.totalAssuntos")
vars_fac <- c("siglaTribunal", "grau", "dadosBasicos.orgaoJulgador.instancia")
vars_str <- c("descricao")
vars_desc <- c("codigoAssunto", "codigoPaiNacional", "descricao", "dadosBasicos.nivelSigilo", "dadosBasicos.codigoLocalidade",
               "dadosBasicos.orgaoJulgador.codigoMunicipioIBGE", "assuntoLocal.codigoAssunto", "assuntoLocal.codigoPaiNacional",
               "assuntoLocal.descricao", "assuntoLocal", "dadosBasicos.tamanhoProcesso", "assuntoLocalPai")

brasil_db_tratados <- brasil_dados_basicos %>% 
  mutate(dadosBasicos.dataAjuizamento = as.POSIXct(dadosBasicos.dataAjuizamento, format = "%Y%m%d%H%M%S"), # transformando em datetime
         across(all_of(vars_int), as.integer),
         across(all_of(vars_fac), as.factor),
         across(all_of(vars_str), as.character)) %>% select(-any_of(vars_desc))
```

```{r}
str(brasil_db_tratados)
```

```{r}
# saveRDS(brasil_db_tratados, "brasil_db_tratados.rds")
```

## Movimentações

```{r}
str(brasil_movimentos)
```

```{r missings_mv}
names(brasil_movimentos[3:4]) # "tipoResponsavelMovimento" "dataHora" 
aggr(brasil_movimentos[3:4], numbers = TRUE) # tipoResponsavelMovimento cerca de 60% mising, dataHora ok

names(brasil_movimentos[5:7]) # "tipoDecisao" "movimentoNacional.codigoNacional" "movimentoLocal.codigoMovimento"  
aggr(brasil_movimentos[5:7], numbers = TRUE) # movimentoNacional.codigoNacional ok, resto mais de 80% mising

names(brasil_movimentos[8:10]) # "movimentoLocal.codigoPaiNacional"  "orgaoJulgador.nomeOrgao" "orgaoJulgador.codigoMunicipioIBGE"
aggr(brasil_movimentos[8:10], numbers = TRUE) # codigoPaiNacional praticamente tudo missing, resto cerca de 60%

names(brasil_movimentos[11:13]) # "orgaoJulgador.codigoOrgao" "orgaoJulgador.instancia" "millisInsercao"    
aggr(brasil_movimentos[11:13], numbers = TRUE) # millisInsercao ok, resto cerca de 60% missing

names(brasil_movimentos[14:16]) # "identificadorMovimento" "movimentoLocal" "nivelSigilo"
aggr(brasil_movimentos[14:16], numbers = TRUE) # identificadorMovimento ok, movimentoLocal tudo missing, nivelSigilo ok mas não relevante (nenhum sigilioso)

names(brasil_movimentos[17:19]) # "orgaoJulgador" "movimentoLocal.movimentoLocalPai" "movimentoNacional" 
aggr(brasil_movimentos[17:19], numbers = TRUE) # tudo missing
```

```{r transformacao_mv}
vars_int <- c()
vars_fac <- c("siglaTribunal", "orgaoJulgador.instancia")
vars_str <- c()
vars_desc <- c("tipoResponsavelMovimento", "tipoDecisao", "movimentoLocal.codigoMovimento", "movimentoLocal.codigoPaiNacional", 
               "orgaoJulgador.codigoMunicipioIBGE", "movimentoLocal", "nivelSigilo", "orgaoJulgador", "movimentoLocal.movimentoLocalPai", "movimentoNacional")

brasil_mov_tratados <- brasil_movimentos %>% 
  mutate(dataHora = as.POSIXct(dataHora, format = "%Y%m%d%H%M%S"), # transformando em datetime
         across(all_of(vars_int), as.integer),
         across(all_of(vars_fac), as.factor),
         across(all_of(vars_str), as.character)) %>% select(-any_of(vars_desc))
```

```{r}
str(brasil_mov_tratados)
```

```{r}
# saveRDS(brasil_mov_tratados, "brasil_mov_tratados.rds")
```
